# Manually form the response string