@dom96 2021-10-29
Very epic. Benchmarked on my machine, improves QPS from 870k to 1.2million!